// utils/types.ts
export type Database = {
// ...
    public: {
        Tables: {
            articles: {
                Row: {
                    article_name: string
                    average_read: number | null
                    category_id: number
                    content: string
                    created_at: string
                    id_article: number
                    intro: string | null
                    picture_link: string
                }
                Insert: {
                    article_name: string
                    average_read?: number | null
                    category_id: number
                    content: string
                    created_at?: string
                    id_article?: number
                    intro?: string | null
                    picture_link: string
                }
                Update: {
                    article_name?: string
                    average_read?: number | null
                    category_id?: number
                    content?: string
                    created_at?: string
                    id_article?: number
                    intro?: string | null
                    picture_link?: string
                }
                Relationships: [
                    {
                        foreignKeyName: "articles_category_id_fkey"
                        columns: ["category_id"]
                        isOneToOne: false
                        referencedRelation: "categories"
                        referencedColumns: ["category_id"]
                    },
                ]
            }
// ...
            categories: {
                Row: {
                    category_id: number
                    category_name: string
                }
                Insert: {
                    category_id?: number
                    category_name: string
                }
                Update: {
                    category_id?: number
                    category_name?: string
                }
                Relationships: []
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
    PublicTableNameOrOptions extends
            | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
        | { schema: keyof Database },
    TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
        ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
            Database[PublicTableNameOrOptions["schema"]]["Views"])
        : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
    ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
            Row: infer R
        }
        ? R
        : never
    : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
            PublicSchema["Views"])
        ? (PublicSchema["Tables"] &
            PublicSchema["Views"])[PublicTableNameOrOptions] extends {
                Row: infer R
            }
            ? R
            : never
        : never